desloppify: move field aliases to leaf module
