+++ src/lib.rs
@@ -16,6 +16,7 @@ pub mod commands;
 pub mod config;
 pub mod credentials;
 pub mod error;
+pub(crate) mod field_aliases;
 pub(crate) mod http;
 #[expect(clippy::expect_used)]
 pub mod output;
